[Q20-3] - Separated custom Authentication provider from Citesphere connector
